Move handleException method to base command
