+++ packages/publications/src/Commands/ValidatingCommand.php
@@ -69,29 +69,6 @@ class ValidatingCommand extends Command
         return $selection;
     }
 
-    /**
-     * Handle an exception that occurred during command execution.
-     *
-     * @param  string|null  $file  The file where the exception occurred. Leave null to auto-detect.
-     * @return int The exit code
-     */
-    public function handleException(Exception $exception, ?string $file = null, ?int $line = null): int
-    {
-        // When testing it might be more useful to see the full stack trace, so we have an option to actually throw the exception.
-        if (config('app.throw_on_console_exception', false)) {
-            throw $exception;
-        }
-
-        // If the exception was thrown from the same file as a command, then we don't need to show which file it was thrown from.
-        if (str_ends_with($file ?? $exception->getFile(), 'Command.php')) {
-            $this->error("Error: {$exception->getMessage()}");
-        } else {
-            $this->error(sprintf('Error: %s at ', $exception->getMessage()).sprintf('%s:%s', $file ?? $exception->getFile(), $line ?? $exception->getLine()));
-        }
-
-        return Command::FAILURE;
-    }
-
     protected function translate(string $name, string $error): string
     {
         return __($error, [
